Personal-work/secure-vault/requirements.txt
flask
pymongo
bcrypt
pycryptodome
cryptography
